arkSoft_phishing_analysis/requirements.txt
python-Levenshtein==0.27.1
beautifulsoup4==4.13.3
urllib3==1.26.20